Refactor processing into separate singleton class.
